move elevator constants to dedicated class
