+++ src/main/java/frc/robot/auto/BlueL4AlgaeAuto.java
@@ -7,12 +7,13 @@ import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.Constants.LifterConstants.LifterState;
 import frc.robot.drivetrain.Drivetrain;
 import frc.robot.elevator.AlgaeRoller;
 import frc.robot.elevator.CoralRoller;
 import frc.robot.elevator.Elevator;
 import frc.robot.elevator.Lifter;
+import frc.robot.elevator.ElevatorConstants.LifterConstants.LifterState;
+
 import java.util.Optional;
 
 public class BlueL4AlgaeAuto extends AutoMode {
