+++ src/main/java/frc/robot/elevator/AlgaeWrist.java
@@ -15,8 +15,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.FunctionalCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants.AlgaeWristConstants;
-import frc.robot.Constants.AlgaeWristConstants.AlgaeWristState;
+import frc.robot.elevator.ElevatorConstants.AlgaeWristConstants;
+import frc.robot.elevator.ElevatorConstants.AlgaeWristConstants.AlgaeWristState;
 import frc.robot.Constants.RobotConstants;
 
 public class AlgaeWrist extends SubsystemBase {
